Make `SproxydClient` constructor arguments explicit
Previously the configuration of an `SproxydClient` instance (e.g. host,
path or timeout settings) was passed in through the `conf` argument
which was used as a stringly-typed dictionary. Furthermore, some values
were interpreted from their string value.

This commit makes the configuration arguments explicit, and no longer
allows string arguments which need to be parsed somehow.