Disregard empty sproxyd endpoint
The `SproxydClient` constructor accepts an iterable of strings as
the `endpoints` parameter. Problem is when that `endpoints` parameter
comes from the parsing of an `oslo_config.ListOpt` and the user
adds some trailing comma then `endpoints` could be `['http://blag', '']`.

In that case the contructor raises a weird  "ValueError: Unknown endpoint
scheme: '' in ''". We shouldn't fail on something as simple as a trailing
comma. Instead just ignore it.